Config file format & interpretation overhaul.
Starting from this branch, the configuration file format and processing
routines are completely rewritten.  ConfigParser is no longer used.
Instead, we use our own simple key-value pair reader to process the
config file.  Semantically this should be backward compatible except for
possibly altered key names (such as "address" -> "server").

The new format is very simple:
1. Valid line contains key-value pairs in the format of
    key = value
with possibly arbitrary whitespace characters surrounding the key, the
equality sign, or the value;
2. The key or the value can contain whitespace characters as long as
it's not the first or last character;
3. Valid key-value pair must contain non-empty key and non-empty value;
4. Inline comments starts with a hash symbol (#) and are ignored;
5. Any line not in the valid key-value pair format are treated as
comment and is ignored.

Therefore, the presence of "sections" in the new config format is purely
superficial and is only intended for backward compatibility.  It can be
safely removed and no section information is used internally anywhere.

The ultimate goal is to unify the config file and command line options.

The removal of the clumsy ConfigParser module also reduces the number of
internal calls to the re package.