Enable WIN32 (type=2) login mode
On Mar 1, 2013 the type=3 login mode was disabled by BNU's server.  To
overcome this, we implement the password and MAC address encoding
scheme used in the type=2 login mode and switch to that by default.
The legacy mode can be enabled by the "-e" switch at the command line.

In this mode, the server calculates the encoded password using its
current UNIX time.  Supposedly this timestamp is shared by the client
and the server yet unknown to a 3rd party.  The credentials are still
passed in plain HTTP packets.  Therefore it is now exactly as insecure
as it was before.

The server is unable to distinguish a wrong password from a right one
wrongly encoded due to time difference of the client and the server.
Therefore, syncing is needed, and this is done by reading the timestamp
appended to the "password_error" message.  If the message still appear
after syncing, we know the password is wrong.