Replace pid file with permanent named pipe
To avoid possible pid collision (i.e. after the daemon crashes),
a new method based on named fifo pipe is used to control the daemons.
The program will mkfifo() a file at ~/.pysrun.fifo if none is found.
This fifo will be reused by future sessions.  In the future, the path
to this file could be made configurable.